requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.22
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3
bcrypt>=4.0
